requirements.txt
Flask==3.0.2
requests==2.25.1
Flask-Session==0.4.0
openai==0.27.4
python-dotenv